resolve circular dependencies
